+++ app_plus/src/main/java/org/gkisalatiga/plus/screen/ScreenPersembahan.kt
@@ -12,6 +12,8 @@ package org.gkisalatiga.plus.screen
 
 import android.app.Application
 import android.content.ClipData
+import android.content.Context
+import android.content.Intent
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.BackHandler
@@ -31,7 +33,9 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.automirrored.filled.Help
+import androidx.compose.material.icons.filled.Share
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
@@ -57,14 +61,23 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.core.content.FileProvider
 import coil.compose.AsyncImage
 import org.gkisalatiga.plus.R
 import org.gkisalatiga.plus.composable.TopAppBarColorScheme
 import org.gkisalatiga.plus.data.ActivityData
 import org.gkisalatiga.plus.db.MainCompanion
 import org.gkisalatiga.plus.lib.AppNavigation
+import org.gkisalatiga.plus.lib.Logger
+import org.gkisalatiga.plus.lib.LoggerType
 import org.gkisalatiga.plus.services.ClipManager
+import org.gkisalatiga.plus.services.InternalFileManager
 import org.json.JSONObject
+import java.io.File
+import java.io.FileOutputStream
+import java.net.URL
+import java.util.concurrent.Executors
+
 
 class ScreenPersembahan (private val current : ActivityData) : ComponentActivity() {
     private var isFirstElement = false
@@ -130,6 +143,14 @@ class ScreenPersembahan (private val current : ActivityData) : ComponentActivity
                         modifier = Modifier.fillMaxWidth(),
                         contentScale = ContentScale.FillWidth
                     )
+                }
+            }
+
+            // Display the QRIS share button.
+            Button(onClick = { ScreenPersembahanCompanion.shareQrisImage(current.ctx) }) {
+                Row (verticalAlignment = Alignment.CenterVertically, horizontalArrangement = Arrangement.Center) {
+                    Icon(Icons.Default.Share, "Share icon", tint = Color.White, modifier = Modifier.padding(end = 10.dp))
+                    Text(stringResource(R.string.offertory_qris_share_btn_text), color = Color.White)
                 }
             }
 
@@ -293,5 +314,64 @@ class ScreenPersembahanCompanion : Application() {
 
         /* The screen's remembered scroll state. */
         var rememberedScrollState: ScrollState? = null
+
+        /* The QRIS location of the image in the phone's storage. */
+        var absolutePathToQrisFile: String = String()
+        val savedFilename = "QRIS GKI Salatiga.png"  // --- allows space and uppercase letters in the file name because it is the most user-friendly file export naming.
+
+        /* Downloading the QRIS image and save/share it outside the app. */
+        fun shareQrisImage(ctx: Context) {
+            val executor = Executors.newSingleThreadExecutor()
+            executor.execute {
+                try {
+                    // Downloading the data.
+                    val streamIn = URL(OFFERTORY_QRIS_SOURCE).openStream()
+                    val decodedData: ByteArray = streamIn.readBytes()
+
+                    // Writing into the file.
+                    val outFile = File(absolutePathToQrisFile)
+                    FileOutputStream(outFile).let {
+                        it.flush()
+                        it.write(decodedData)
+                        it.close()
+                        it
+                    }
+
+                    // The URI to the QRIS image provided by FileProvider.
+                    val authority = InternalFileManager(ctx).FILE_PROVIDER_AUTHORITY
+                    val secureUri = FileProvider.getUriForFile(ctx, authority, outFile)
+
+                    // The strings to be attached to the intent.
+                    val qrisShareSheetTitle = ctx.getString(R.string.offertory_qris_sharesheet_title)
+                    val qrisShareSheetText = ctx.getString(R.string.offertory_qris_sharesheet_text)
+
+                    // Creating the intent.
+                    val shareIntent = Intent().apply {
+                        action = Intent.ACTION_SEND
+                        putExtra(Intent.EXTRA_STREAM, secureUri)
+                        putExtra(Intent.EXTRA_TEXT, qrisShareSheetText)
+                        putExtra(Intent.EXTRA_TITLE, qrisShareSheetTitle)  // --- this is not currently successful at setting the chooser title.
+                        addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+                        type = "image/png"
+
+                        // Attach clipboard.
+                        // This prevents "Writing exception to parcel" (java.lang.SecurityException) when attempting to load thumbnails.
+                        // The error in itself is not harmful; the images can still be shared or saved to the destination file.
+                        // SOURCE: https://stackoverflow.com/a/64541741
+                        clipData = ClipData(
+                            "QRIS Image of GKI Salatiga",
+                            listOf("image/png").toTypedArray(),
+                            ClipData.Item(secureUri)
+                        )
+                    }
+
+                    // Sharing the data.
+                    ctx.startActivity(Intent.createChooser(shareIntent, null))
+                } catch (e: Exception) {
+                    Logger.logTest({}, "Exception: ${e.message}", LoggerType.ERROR)
+                    e.printStackTrace()
+                }
+            }
+        }
     }
 }